# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blocker X - Discord Bot Hosting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/discord/callback:
    get:
      operationId: discordCallback
      tags: [auth]
      summary: Discord OAuth2 callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to dashboard or invite page
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Invite codes
...
  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List user's bots
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"

    post:
      operationId: createBot
      tags: [bots]
      summary: Create a new bot project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{botId}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get a bot by ID
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

    patch:
      operationId: updateBot
      tags: [bots]
      summary: Update a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotUpdate"
      responses:
        "200":
          description: Updated bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /bots/{botId}/start:
    post:
      operationId: startBot
      tags: [bots]
      summary: Start a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot starting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /bots/{botId}/stop:
    post:
      operationId: stopBot
      tags: [bots]
      summary: Stop a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /bots/{botId}/restart:
    post:
      operationId: restartBot
      tags: [bots]
      summary: Restart a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot restarting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /bots/{botId}/status:
    get:
      operationId: getBotStatus
      tags: [bots]
      summary: Get bot runtime status
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"

  # Deployments
...
  /bots/{botId}/deploy:
    post:
      operationId: deployBot
      tags: [deployments]
      summary: Deploy a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{deploymentId}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment details
      parameters:
        - name: deploymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"

  # File Manager
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
        discordId:
          type: string
        username:
          type: string
        discriminator:
          type: string
        avatar:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, premium]
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        hasInvite:
          type: boolean
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
      required:
        - id
        - discordId
        - username
        - plan
        - isAdmin
        - isBanned
        - hasInvite
        - createdAt
...
    Bot:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
          enum: [python, javascript]
        status:
          type: string
          enum: [stopped, starting, running, errored, deploying]
        userId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        mainFile:
          type: ["string", "null"]
        r2Prefix:
          type: string
      required:
        - id
        - name
        - language
        - status
        - userId
        - createdAt
        - updatedAt
        - r2Prefix
    BotInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string
          enum: [python, javascript]
        mainFile:
          type: string
      required:
        - name
        - language
    BotUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        mainFile:
          type: string
    BotStatus:
      type: object
      properties:
        botId:
          type: string
        status:
          type: string
        uptime:
          type: ["integer", "null"]
        memoryMB:
          type: ["number", "null"]
        lastStarted:
          type: ["string", "null"]
        lastStopped:
          type: ["string", "null"]
      required:
        - botId
        - status
    Deployment:
      type: object
      properties:
        id:
          type: string
        botId:
          type: string
        botName:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, cancelled]
        startedAt:
          type: string
        finishedAt:
          type: ["string", "null"]
        logs:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
      required:
        - id
        - botId
        - botName
        - userId
        - status
        - startedAt
...