# lib/api-zod/src/generated/types/botInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Blocker X - Discord Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BotInputLanguage } from './botInputLanguage';

export interface BotInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  language: BotInputLanguage;
  mainFile?: string;
}